requirements.txt
# Python Version Requirement
# Python 3.11 is required for this project
# Python 3.13 is NOT compatible due to SQLAlchemy typing issues

# Web Framework
Flask==2.3.3
Werkzeug==2.3.7
Jinja2==3.1.6
itsdangerous==2.2.0
blinker==1.9.0
click==8.1.8
MarkupSafe==3.0.2
Flask-WTF==1.2.1
WTForms==3.2.1
gunicorn==21.2.0
whitenoise==6.6.0  # For serving static files

# Database
# Note: SQLAlchemy 2.0.23 is not compatible with Python 3.13
# due to typing issues (TypingOnly class inheritance restrictions)
SQLAlchemy==2.0.23
greenlet==3.1.1
PyMySQL==1.1.0
psycopg2-binary==2.9.9

# File Processing
Pillow==10.0.1
ExifRead==3.0.0
PyPDF2==3.0.1
python-docx==1.0.1
openpyxl==3.1.2
et_xmlfile==2.0.0
lxml==5.3.1
python-magic==0.4.27

# Media Metadata Extraction
hachoir==3.2.0
mutagen==1.47.0
pymediainfo==6.0.1
pikepdf==8.11.2
oletools==0.60.1

# Image Processing
opencv-python==4.11.0.86
numpy==2.2.4
scipy==1.15.2

# Utilities
python-dotenv==1.0.0
typing_extensions==4.12.2
packaging==24.2

# AI Integration
google-generativeai==0.8.5


# Testing
pytest==7.4.2
pluggy==1.5.0
iniconfig==2.1.0
exceptiongroup==1.2.2
tomli==2.2.1
